Enchanting ITab and BillStack
I wasn't sure how this worked, so I wound up following the framework from the IBillGiver interface to the RecipeDefs before recalling that the ITab is the first place I should have worked on.  I only managed to get a little work done on the ITab because of time constraints.