Ability/Magic proof of concept
TES has so many spells that the player will need to be able to select which spells to have gizmos for.  There will not be room for all of them.  In addition, custom spells will eventually be able to be made at runtime, so that functionality has to be supported now.